CS-6308: Fix duplication
